% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanAQRBABDaily.R
\name{cleanAQRBABDaily}
\alias{cleanAQRBABDaily}
\title{Reads in, cleans, and subdivides Daily BAB data set in data folder.
Currently disabled, Java heap runs out of memory when reading xlsx.}
\usage{
cleanAQRBABDaily()
}
\description{
Reads in, cleans, and subdivides Daily BAB data set in data folder.
Currently disabled, Java heap runs out of memory when reading xlsx.
}

