% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanAQRQMJDaily.R
\name{cleanAQRQMJDaily}
\alias{cleanAQRQMJDaily}
\title{Reads in, cleans, and subdivides daily QMJ data set in data folder.
Currently disabled, Java heap runs out of memory when reading xlsx.}
\usage{
cleanAQRQMJDaily()
}
\description{
Reads in, cleans, and subdivides daily QMJ data set in data folder.
Currently disabled, Java heap runs out of memory when reading xlsx.
}

