% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanAQRQMJMonthly.R
\name{cleanAQRQMJMonthly}
\alias{cleanAQRQMJMonthly}
\title{Reads in, cleans, and subdivides monthly QMJ data set in data folder.
Note that all overlapping AQR data is taken from this dataset.}
\usage{
cleanAQRQMJMonthly()
}
\description{
Reads in, cleans, and subdivides monthly QMJ data set in data folder.
Note that all overlapping AQR data is taken from this dataset.
}

