% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanFFCMA.R
\name{cleanFFCMA}
\alias{cleanFFCMA}
\title{Reads in, cleans, and subdivides FFCMA data set located in data folder.}
\usage{
cleanFFCMA()
}
\description{
Reads in, cleans, and subdivides FFCMA data set located in data folder.
}

