% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanFFHMLDaily.R
\name{cleanFFHMLDaily}
\alias{cleanFFHMLDaily}
\title{Reads in, cleans, and subdivides daily FFHML data set in data folder.}
\usage{
cleanFFHMLDaily()
}
\description{
Reads in, cleans, and subdivides daily FFHML data set in data folder.
}

