% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanFFRMW.R
\name{cleanFFRMW}
\alias{cleanFFRMW}
\title{Reads in, cleans, and subdivides FFRMW data set in data folder.}
\usage{
cleanFFRMW()
}
\description{
Reads in, cleans, and subdivides FFRMW data set in data folder.
}

