% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanFFSMBDaily.R
\name{cleanFFSMBDaily}
\alias{cleanFFSMBDaily}
\title{Reads in, cleans, and subdivides daily FFSMB data set in data folder.}
\usage{
cleanFFSMBDaily()
}
\description{
Reads in, cleans, and subdivides daily FFSMB data set in data folder.
}

