% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanFFSMBMonthly.R
\name{cleanFFSMBMonthly}
\alias{cleanFFSMBMonthly}
\title{Reads in, cleans, and subdivides monthly FFSMB data set in data folder.}
\usage{
cleanFFSMBMonthly()
}
\description{
Reads in, cleans, and subdivides monthly FFSMB data set in data folder.
}

