% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cleanSubFF.R
\name{cleanSubFF}
\alias{cleanSubFF}
\title{Cleans separate Fama-French data frames and adds them to global environment after
splitting the set into thirds, quintiles and deciles.}
\usage{
cleanSubFF(data, spr = TRUE)
}
\arguments{
\item{data}{A data set.}

\item{spr}{Calculates spreads for relevant data by default.}
}
\description{
Cleans separate Fama-French data frames and adds them to global environment after
splitting the set into thirds, quintiles and deciles.
}

