% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetchDf.R
\name{fetchDf}
\alias{fetchDf}
\title{Returns the specific data frame requested.
See ?FFAQR for further information.}
\usage{
fetchDf(name)
}
\arguments{
\item{name}{Name of data frame requested.}
}
\description{
Returns the specific data frame requested.
See ?FFAQR for further information.
}

