% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/findBreak.R
\name{findBreak}
\alias{findBreak}
\title{Finds the row in a data frame at which column 1 is a given string.}
\usage{
findBreak(sheet, start = 1, breakpt)
}
\arguments{
\item{sheet}{A data table.}

\item{start}{Which row to start looking at, going down.}

\item{breakpt}{Which phrase stops the loop}
}
\value{
The number of the row at which to set a breakpoint.
}
\description{
Finds the row in a data frame at which column 1 is a given string.
}

