% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/numValues.R
\name{numValues}
\alias{numValues}
\title{Iterates through columns to transform all data table values into numbers.}
\usage{
numValues(data)
}
\arguments{
\item{data}{A data table.}
}
\value{
The same data set, with all values coerced to numeric.
}
\description{
Iterates through columns to transform all data table values into numbers.
}

