% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/updateData.R
\name{updateData}
\alias{updateData}
\title{Updates all data.}
\usage{
updateData(which = "ALL")
}
\arguments{
\item{which}{Specifies which data sets to load. Defaults to all, "AQR" and "FF" load only respective data sets.}
}
\description{
Updates all data.
}

